// dbschema/interfaces.ts
export namespace std {
  export interface BaseObject {
    "id": string;
  }
  export interface $Object extends BaseObject {}
  export interface FreeObject extends BaseObject {}
  export type JsonEmpty = "ReturnEmpty" | "ReturnTarget" | "Error" | "UseNull" | "DeleteKey";
}
// ...
export interface Questions extends std.$Object {
  "answer": number;
  "description": string;
  "question": string;
  "name"?: string | null;
}
export interface Task extends std.$Object {
  "description"?: string | null;
  "name": string;
  "questions"?: unknown | null;
}
export interface User extends std.$Object {
  "tasks": unknown;
  "email"?: string | null;
  "name"?: string | null;
}
export interface test0 extends std.$Object {
  "hasasdda": unknown;
}
// ...
export namespace testModule {
  export interface avc extends std.$Object {
    "name": string;
    "question": string;
    "answer"?: number | null;
    "description"?: string | null;
  }
  export interface hasasdda extends std.$Object {
    "users": unknown;
  }
}